+++ fili-core/src/main/java/com/yahoo/bard/webservice/application/DruidDimensionsLoader.java
@@ -169,6 +169,9 @@ public class DruidDimensionsLoader extends Loader<Boolean> {
                         }
                     }
                 }
+
+                // Tell the dimension it's been updated
+                dimension.setLastUpdated(DateTime.now());
             }
         };
     }