+++ CHANGELOG.md
@@ -31,6 +31,9 @@ Current
 
 ### Fixed:
 
+- [`DruidDimensionsLoader` doesn't set the dimension's lastUpdated date](https://github.com/yahoo/fili/pull/24)
+  * `DruidDimensionsLoader` now properly sets the `lastUpdated` field after it finished processing the Druid response
+
 
 ### Known Issues: